+++ lib/row_row_row.dart
@@ -1,3 +1,4 @@
-int calculate() {
-  return 6 * 7;
-}
+/// Exports the main generation function.
+library row_row_row;
+
+export 'src/generator.dart' show generate;