+++ test/row_row_row_test.dart
@@ -2,7 +2,5 @@ import 'package:row_row_row/row_row_row.dart';
 import 'package:test/test.dart';
 
 void main() {
-  test('calculate', () {
-    expect(calculate(), 42);
-  });
+  // TODO make tests
 }